Move FullNode::new_window into window module
